refactor: shared 폴더의 sidebar 컴포넌트 수정
