+++ src/shared/ui/sidebar/index.ts
@@ -1,0 +1,5 @@
+export { default as SidebarMenuFooter } from './sidebar-menu-footer'
+export { default as SidebarMenuHeader } from './sidebar-menu-header'
+export { default as SidebarMenuList } from './sidebar-menu-list'
+export { default as SidebarMenu } from './sidebar-menu'
+export { default as Sidebar } from './sidebar'
