+++ src/shared/ui/sidebar/sidebar.tsx
@@ -1,9 +1,8 @@
 import { motion } from 'framer-motion'
 
-import { useGNBContext } from './gnb-conext'
-import SideBarMenu from './sidebar-menu'
-import SideBarMenuFooter from './sidebar-menu-footer'
-import SideBarMenuHeader from './sidebar-menu-header'
+import { useGNBContext } from '../gnb/gnb-conext'
+
+import type { ReactNode } from 'react'
 
 /**
  * @description 햄버거 버튼 클릭시 나타나는 사이드바 입니다
@@ -14,8 +13,13 @@ const sidebarVariants = {
   open: { x: 0 },
 }
 
-export default function SideBar() {
+interface Props {
+  children: ReactNode
+}
+
+export default function SideBar({ children }: Props) {
   const { isOpen } = useGNBContext()
+
   return (
     <motion.nav
       initial="closed"
@@ -25,11 +29,7 @@ export default function SideBar() {
       transition={{ type: 'tween', duration: 0.3 }}
       className="w-[16rem] h-full fixed top-0 right-0 flex flex-col justify-between bg-white p-[0.625rem] z-side-bar rounded-s-lg"
     >
-      <div>
-        <SideBarMenuHeader username="이름" />
-        <SideBarMenu />
-      </div>
-      <SideBarMenuFooter />
+      {children}
     </motion.nav>
   )
 }
